Playback Application
After running the application, the result will be something like this:
rice. 17.2. Even though you can click the buttons, there is no response.
occurs because the program does not contain any methods intended
to receive user data and respond to it. Additional
information on this topic will be given in session 19.